feat!: restructure runtime functions and related data
Based on issue #17, all runtime related data has been moved to the main file. Aside from that there are a lot more comments now
